+++ bugsnag-android-core/src/main/java/com/bugsnag/android/internal/ImmutableConfig.kt
@@ -21,10 +21,10 @@ import com.bugsnag.android.NoopLogger
 import com.bugsnag.android.Session
 import com.bugsnag.android.Telemetry
 import com.bugsnag.android.ThreadSendPolicy
+import com.bugsnag.android.anyCauseMatches
 import com.bugsnag.android.errorApiHeaders
 import com.bugsnag.android.internal.dag.Provider
 import com.bugsnag.android.internal.dag.ValueProvider
-import com.bugsnag.android.safeUnrollCauses
 import com.bugsnag.android.sessionApiHeaders
 import java.io.File
 import java.util.regex.Pattern
@@ -132,7 +132,7 @@ data class ImmutableConfig(
      */
     @VisibleForTesting
     internal fun shouldDiscardByErrorClass(exc: Throwable): Boolean {
-        return exc.safeUnrollCauses().any { throwable ->
+        return exc.anyCauseMatches { throwable ->
             val errorClass = throwable.javaClass.name
             shouldDiscardByErrorClass(errorClass)
         }