+++ bugsnag-plugin-android-ndk/src/main/jni/utils/serializer/event_writer.c
@@ -309,18 +309,43 @@ static bool bsg_write_severity_reason(BSG_KSJSONEncodeContext *json,
 
   CHECKED(bsg_ksjsonbeginObject(json, "severityReason"));
   {
+    bsg_error *error = &event->error;
+
     // unhandled == false always means that the state has been overridden by the
     // user, as this codepath is only executed for unhandled native errors
     CHECKED(bsg_ksjsonaddBooleanElement(json, "unhandledOverridden",
                                         !event->unhandled));
-    CHECKED(JSON_CONSTANT_ELEMENT("type", "signal"));
 
-    bsg_error *error = &event->error;
-    CHECKED(bsg_ksjsonbeginObject(json, "attributes"));
-    { CHECKED(JSON_LIMITED_STRING_ELEMENT("signalType", error->errorClass)); }
-    CHECKED(bsg_ksjsonendContainer(json));
+    if (error->type[0] == 'a' && error->type[1] == 'n' &&
+        error->type[2] == 'd' && error->type[3] == 'r' &&
+        error->type[4] == 'o' && error->type[5] == 'i' &&
+        error->type[6] == 'd' && error->type[7] == 0) {
+      CHECKED(JSON_CONSTANT_ELEMENT("type", "unhandledException"));
+    } else {
+      CHECKED(JSON_CONSTANT_ELEMENT("type", "signal"));
+
+      CHECKED(bsg_ksjsonbeginObject(json, "attributes"));
+      { CHECKED(JSON_LIMITED_STRING_ELEMENT("signalType", error->errorClass)); }
+      CHECKED(bsg_ksjsonendContainer(json));
+    }
   }
   CHECKED(bsg_ksjsonendContainer(json));
+  return true;
+error:
+  return false;
+}
+
+static bool bsg_write_java_stackframe(BSG_KSJSONEncodeContext *json,
+                                      bugsnag_stackframe *frame) {
+  CHECKED(bsg_ksjsonbeginObject(json, NULL));
+  {
+    CHECKED(JSON_LIMITED_STRING_ELEMENT("method", frame->method));
+    CHECKED(JSON_LIMITED_STRING_ELEMENT("file", frame->filename));
+    CHECKED(
+        bsg_ksjsonaddIntegerElement(json, "lineNumber", frame->line_number));
+  }
+  CHECKED(bsg_ksjsonendContainer(json));
+
   return true;
 error:
   return false;
@@ -377,11 +402,17 @@ error:
 
 static bool bsg_write_stacktrace(BSG_KSJSONEncodeContext *json,
                                  bugsnag_stackframe *stacktrace,
-                                 size_t frame_count) {
+                                 size_t frame_count, bool is_native_error) {
 
   for (int findex = 0; findex < frame_count; findex++) {
-    if (!bsg_write_stackframe(json, &stacktrace[findex], findex == 0)) {
-      goto error;
+    if (is_native_error) {
+      if (!bsg_write_stackframe(json, &stacktrace[findex], findex == 0)) {
+        goto error;
+      }
+    } else {
+      if (!bsg_write_java_stackframe(json, &stacktrace[findex])) {
+        goto error;
+      }
     }
   }
 
@@ -391,13 +422,16 @@ error:
 }
 
 static bool bsg_write_error(BSG_KSJSONEncodeContext *json, bsg_error *error) {
+  bool is_native_error =
+      strncmp(error->type, "android", sizeof(error->type)) != 0;
+
   CHECKED(bsg_ksjsonbeginArray(json, "exceptions"));
   {
     CHECKED(bsg_ksjsonbeginObject(json, NULL));
     {
       CHECKED(JSON_LIMITED_STRING_ELEMENT("errorClass", error->errorClass));
       CHECKED(JSON_LIMITED_STRING_ELEMENT("message", error->errorMessage));
-      CHECKED(JSON_CONSTANT_ELEMENT("type", "c"));
+      CHECKED(JSON_LIMITED_STRING_ELEMENT("type", error->type));
 
       const ssize_t frame_count = error->frame_count;
       // assuming that the initial frame is the program counter. This logic will
@@ -407,7 +441,8 @@ static bool bsg_write_error(BSG_KSJSONEncodeContext *json, bsg_error *error) {
       if (frame_count > 0) {
         CHECKED(bsg_ksjsonbeginArray(json, "stacktrace"));
         {
-          if (!bsg_write_stacktrace(json, error->stacktrace, frame_count)) {
+          if (!bsg_write_stacktrace(json, error->stacktrace, frame_count,
+                                    is_native_error)) {
             goto error;
           }
         }