+++ features/steps/android_steps.rb
@@ -29,10 +29,8 @@ def execute_command(action, scenario_name = '')
 end
 
 def press_at(x, y)
-  touch_action = Appium::TouchAction.new
-  touch_action.press({x: x, y: y}).wait(1).release
   begin
-    touch_action.perform
+    Maze::Api::Appium::UiManager.new.touch_at(x, y)
   rescue Selenium::WebDriver::Error::ServerError
     # Just ignore it, the press still seems to work
   end
@@ -125,6 +123,7 @@ When("I tap the screen {int} times") do |count|
   (1..count).each { |i|
     begin
       press_at 500, 300
+      sleep 1
     rescue Selenium::WebDriver::Error::ElementNotInteractableError, Selenium::WebDriver::Error::InvalidElementStateError
       # Ignore it
     end